app/requirements.txt
absl-py==2.0.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
automodel==0.5.9
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
bqplot==0.12.36
branca==0.6.0
Brotli==1.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colour==0.1.5
comm==0.1.2
compress==0.0.3
contourpy==1.0.7
cors==1.0.1
cycler==0.11.0
DateTime==5.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
earthengine-api==0.1.344
ee-extra==0.0.15
eerepr==0.0.4
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
filelock==3.9.1
Flask==2.3.2
Flask-Compress==1.14
Flask-Cors==4.0.0
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.39.0
fqdn==1.5.1
fsspec==2023.9.2
future==0.18.3
gast==0.4.0
gdown==4.6.4
geeadd==0.5.6
geemap==0.20.1
geocoder==1.38.1
geojson==3.0.1
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
httplib2==0.21.0
huggingface-hub==0.17.2
idna==3.4
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.21.3
ipyleaflet==0.17.2
ipython==8.11.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.7.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonify==0.5
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
logzero==1.7.0
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
morningstar==0.3.1
MyApplication==0.1.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
node==1.2.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
odict==1.9.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
plumber==1.7
plumbum==1.8.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyCRS==1.0.2
pyfunds==1.1.0
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
python-box==7.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
ratelim==0.1.6
regex==2023.8.8
repoze.lru==0.7
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rm==2020.12.3
Routes==2.5.1
rpyc==5.3.1
rsa==4.9
safetensors==0.3.3
sankee==0.2.3
scipy==1.10.1
scooby==0.7.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
sudo==1.0.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.4
tokenizers==0.13.3
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
transformers==4.33.2
typing_extensions==4.5.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
values==2020.12.3
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==3.6.2
windrose==1.8.1
wrapt==1.15.0
xyzservices==2023.2.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==5.0
zope.proxy==5.0.0
